+++ config/settings/base.py
@@ -492,6 +492,19 @@ PROJECTROLES_ALLOW_LOCAL_USERS = env.bool(
 # Warn about unsupported browsers (IE)
 PROJECTROLES_BROWSER_WARNING = True
 
+# Disable default CDN JS/CSS includes to replace with your local files
+PROJECTROLES_DISABLE_CDN_INCLUDES = env.bool(
+    'PROJECTROLES_DISABLE_CDN_INCLUDES', False
+)
+# Paths/URLs to optional global includes to supplement/replace default ones
+PROJECTROLES_CUSTOM_JS_INCLUDES = env.list(
+    'PROJECTROLES_CUSTOM_JS_INCLUDES', None, []
+)
+PROJECTROLES_CUSTOM_CSS_INCLUDES = env.list(
+    'PROJECTROLES_CUSTOM_CSS_INCLUDES', None, []
+)
+
+
 # Timeline app settings
 TIMELINE_PAGINATION = 15